Rename chord/workbench/api/utils/likeAndPlayCount.ts -> src/chord/workbench/api/utils/statistic.ts
