+++ src/chord/workbench/parts/mainView/browser/component/album/album.tsx
@@ -5,7 +5,7 @@ import 'chord/css!../../media/album';
 import * as React from 'react';
 import { connect } from 'react-redux';
 
-import { getLikeAndPlayCount } from 'chord/workbench/api/utils/likeAndPlayCount';
+import { getLikeAndPlayCount } from 'chord/workbench/api/utils/statistic';
 
 import { ESize } from 'chord/music/common/size';
